Add flask-restx to support swagger docs
Flask-restx enables documenting the API through a auto-generated
Swagger from code. The Swagger is available at the root path of the API.

An example of a route/controller and how it can be documented can be
seen at backend/apis/example_namespace.py.

Due to Flask by default seraching for the __init__.py the app.py was
renamed.